fix(qa): wrap auth-unavailable fallback in modal shell + fix strict-mode close loop
Two real bugs caught by exhaustive QA on the auth modal flow:

1. StrictMode double-mount cleanup loop. The shell's open-and-close effect
   called dialog.close() on cleanup, which fired the `close` event, which
   the handler treated as "user dismissed" and called router.back().
   In dev StrictMode this fired between the double-mount cycle, so every
   modal open was immediately reverted back to /. The /signup URL flashed
   then disappeared before the user saw anything. Switched to listening
   for the `cancel` event instead (fires on Escape only, not on programmatic
   close) and removed the cleanup close() call. Dialog cleanup relies on
   DOM removal, which is enough.

2. Clerk-mock fallback rendered in flow, not as a modal. The @auth slot
   layout short-circuited to <AuthUnavailableCard /> without a modal
   wrapper, so dev:local:browse showed the card appended below the
   homepage instead of over it. Prod (real Clerk) worked fine; dev was
   confusing. Extracted AuthModalShell as a shared client component and
   wrapped both branches of the layout's clerk-availability check in it.
   Dev and prod now show the same modal chrome; only the content differs.

Side effect: the intercepted /signup page is simpler — the dialog +
dismissal logic lives in the shared shell, not duplicated in page.tsx.

Verified at localhost:3011: submit → /signup URL persists, dialog
mounts, focus trapped on first button inside dialog, unavailable card
centered in dialog with backdrop blur over the dimmed homepage.
